% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.fut}
\alias{print.fut}
\title{Print basic information about the futility design}
\usage{
\method{print}{fut}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fut}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Print the results of \code{fut} object.
}
\description{
Print the power loss or sample size inflation factor due to
the planned futility analysis.
}
\examples{
# see example for fut
}
\seealso{
\code{\link{fut}}, \code{\link{summary.fut}}
}
\keyword{fut}
