% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.summary.fut}
\alias{print.summary.fut}
\title{Print method for summary.fut objects}
\usage{
\method{print}{summary.fut}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{summary.fut}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Print the detailed summary of the futility design.
}
\seealso{
\code{\link{fut}}, \code{\link{summary.fut}}.
}
